新增blueprint機制整理code
