+++ attraction.py
@@ -1,0 +1,30 @@
+from flask import Blueprint,request
+from data import data_handle
+import json
+from main import db
+
+attraction_app = Blueprint("attraction_app",__name__)
+
+
+@attraction_app.route("/api/attractions")
+def attr():
+	try:
+		page = int(request.args.get("page"))+1
+		name = request.args.get("keyword")
+		if name:
+			ans = data_handle.filter_by_keyword(db,page,name)
+		else:
+			ans = data_handle.filter_by_page(db,page)
+		return ans
+	except:
+		return json.dumps({"error":True,"message":"伺服器內部錯誤"},ensure_ascii=False),500
+
+@attraction_app.route("/api/attraction/<id>")
+def attr2(id):
+	try:
+		res = data_handle.filter_by_id(db,id)
+		if not res:
+			return json.dumps({"error":True,"message":"景點編號錯誤"},ensure_ascii=False),400
+		return res
+	except:
+		return json.dumps({"error":True,"message":"伺服器內部錯誤"},ensure_ascii=False),500
